requirements.txt
flask==3.0.2
pymysql==1.1.2
cryptography==46.0.3
python-dotenv==1.0.1
gunicorn==21.2.0